fix(frontend) fixed 'non-serializable action' error
