+++ webapp/javascript/components/enchaners.tsx
@@ -1,0 +1,22 @@
+import { connect } from 'react-redux';
+import { RootState } from '@pyroscope/redux/store';
+import { selectUIState } from '@pyroscope/redux/reducers/views';
+
+import { setCollapsedUI } from '../redux/actions';
+
+export interface ICollapsible {
+  collapsed: boolean;
+  setCollapsed: (boolean) => void;
+}
+
+export const withCollapsible = (path) =>
+  connect(
+    (state: RootState) => ({
+      collapsed: selectUIState(state)(path),
+    }),
+    (dispatch) => ({
+      setCollapsed: (value) => {
+        dispatch(setCollapsedUI(path, value));
+      },
+    })
+  );
